feat: mirror directory structure for tests
